Added Authentication in add & edit articles

+++ routes/articles.js
@@ -3,9 +3,11 @@ const router = express.Router();
 
 let Article = require('../models/Article');
 
+let User = require('../models/User');
 
 
-router.get('/add', (req, res) => {
+
+router.get('/add', ensureAuthenticated, (req, res) => {
     res.render('add_articles', {
         title:'Add Articles'
     });
@@ -14,7 +16,7 @@ router.get('/add', (req, res) => {
 router.post('/add', (req, res) => {
 
     req.checkBody('title','Title is required').notEmpty();
-    req.checkBody('author','Author is required').notEmpty();
+    // req.checkBody('author','Author is required').notEmpty();
     req.checkBody('body','Body is required').notEmpty();
 
 
@@ -28,7 +30,7 @@ router.post('/add', (req, res) => {
     } else {
         let article = new Article();
         article.title = req.body.title;
-        article.author = req.body.author;
+        article.author = req.user._id;
         article.body = req.body.body;
     
         article.save((err) =>{
@@ -50,8 +52,12 @@ router.post('/add', (req, res) => {
 });
 
 
-router.get('/edit/:id', (req, res) => {
+router.get('/edit/:id', ensureAuthenticated, (req, res) => {
     Article.findById(req.params.id, (err, article) => {
+        if(article.author != req.user._id){
+            req.flash('danger', 'Not Authorized');
+            res.redirect('/');
+        }
         res.render('edit_article', {
             title:'Edit Article',
             article:article
@@ -80,13 +86,24 @@ router.post('/edit/:id', (req, res) => {
 });
 
 router.delete('/:id', (req, res) => {
+
+    if(!req.user._id){
+        res.status(500).send();
+    }
+
     let query = {_id:req.params.id}
 
-    Article.deleteOne(query, (err) => {
-        if(err){
-            console.log(err);
-        }
-        res.send('Success');
+    Article.findById(req.params.id, function(err, article){
+       if(article.author != req.user._id){
+            res.status(500).send();
+       } else {
+            Article.deleteOne(query, (err) => {
+                if(err){
+                    console.log(err);
+                }
+                res.send('Success');
+            });
+       }
     });
 });
 
@@ -94,16 +111,26 @@ router.delete('/:id', (req, res) => {
 
 router.get('/:id', (req, res) => {
     Article.findById(req.params.id, (err, article) => {
-        res.render('article', {
-            article:article
+        User.findById(article.author, function(err, user){
+            res.render('article', {
+                article:article,
+                author: user.name
+            });
         });
-       
         // console.log(article);
         // return;
     });
 });
 
 
+function ensureAuthenticated(req, res, next){
+    if(req.isAuthenticated()){
+        return next();
+    } else {
+       req.flash('danger', 'Please login');
+       res.redirect('/users/login'); 
+    }
+}
